backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1

# Validation / Settings
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.0

# AI (Vertex AI — HIPAA-eligible with GCP BAA)
google-cloud-aiplatform==1.74.0

# Email
resend==2.5.1

# Scheduling
apscheduler==3.10.4

# Templating
jinja2==3.1.5
